fix: reduce duplicate code
