+++ packages/ngx-fast-icon-demo/src/app/comparison/ant-icons.ts
