+++ packages/ngx-fast-icon-demo/src/app/comparison/fast-icon.component.ts
@@ -1,24 +1,19 @@
 import { Component } from '@angular/core';
-import { SUPPORTED_ICONS } from '../../icon-data';
-import { IconTester } from '../../misc/icons-tester/icon-tester.service';
-import { ControllerComponent } from '../../misc/controller.component';
+import { SUPPORTED_ICONS } from '../icon-data';
+import { IconTester } from '../misc/icons-tester/icon-tester.service';
+import { ControllerComponent } from '../misc/controller.component';
 import { AsyncPipe, NgClass } from '@angular/common';
-import { SvgIconComponent } from 'angular-svg-icon';
 import { FastSvgComponent } from '@push-based/ngx-fast-svg';
-
 @Component({
   standalone: true,
   template: `
-    <app-controller [demoLib]='"Angular svg icon"' [tester]='tester' />
+    <app-controller [demoLib]='"Fast svg"' [tester]='tester' />
     <div class='row icons' [ngClass]='tester.layout | async'>
       @for (list of (tester.lists | async); track $index) {
         <ul class='group'>
           @for (icon of list; track $index) {
             <li>
-              <svg-icon
-                [src]='icon'
-                [svgStyle]="{ 'width.px': 30, 'height.px': 30 }"
-              />
+              <fast-svg class='fast-svg-icon' [name]='icon' />
             </li>
           }
         </ul>
@@ -34,16 +29,13 @@ import { FastSvgComponent } from '@push-based/ngx-fast-svg';
   `],
   imports: [
     ControllerComponent,
-    AsyncPipe,
     NgClass,
-    SvgIconComponent,
+    AsyncPipe,
     FastSvgComponent
   ]
 })
-export class AngularComponent {
+export class FastIconRouteComponent {
   constructor(public tester: IconTester) {
-    this.tester.defineSet(
-      SUPPORTED_ICONS.map((i) => `assets/svg-icons/${i}.svg`)
-    );
+    this.tester.defineSet(SUPPORTED_ICONS);
   }
 }
