+++ packages/ngx-fast-icon-demo/src/app/comparison/ionic.component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
-import { SUPPORTED_ICONS } from '../../icon-data';
-import { IconTester } from '../../misc/icons-tester/icon-tester.service';
+import { SUPPORTED_ICONS } from '../icon-data';
+import { IconTester } from '../misc/icons-tester/icon-tester.service';
 import { AsyncPipe, NgClass } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
-import { ControllerComponent } from '../../misc/controller.component';
+import { ControllerComponent } from '../misc/controller.component';
 
 @Component({
   standalone: true,
